requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.1
django-currencies @ git+https://github.com/panosl/django-currencies.git@594f7cd2a2f4e6396565566c739de63599d6a1ce
gunicorn==22.0.0
idna==3.7
jsonfield==2.1.1
packaging==24.1
pillow==10.3.0
python-decouple==3.8
python-dotenv==0.21.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
whitenoise==6.6.0
